+++ src/components/Headers/index.js
@@ -8,7 +8,7 @@ import { makeStyles } from "@material-ui/core/styles";
 import EcoIcon from "@material-ui/icons/Eco";
 
 // account header imports
-import SwipeableTemporaryDrawer from './settings.js';
+import SwipeableTemporaryDrawer from '../../pages/AccountPage/Settings';
 import Grid from "@material-ui/core/Grid";
 
 // change dorm/pw imports
