+++ src/components/Navigation/index.js
@@ -4,8 +4,7 @@ import { Link } from "react-router-dom";
 import * as ROUTES from "../../constants/routes";
 import { AuthUserContext } from "../../services/Session";
 import SignOutButton from "../SignOut";
-
-import styles from "../Navigation.modules.css";
+import "./Navigation.modules.css";
 
 // import your fontawesome library
 import "../FontAwesomeIcons";
